+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### New in Version 6:
 - Read Upgrade
 - Add extensible entity architecture: extend `BaseScheduledCommand` to add custom fields (see README)
+- Fix: Automatically exclude default `ScheduledCommand` entity from Doctrine mappings when custom entity class is configured
 - Ensure that the log file reside in the log directory and finish with a ".log" extension
 - Add command `scheduler:disable` to disable one or all scheduled commands
 - Add support of Symfony 7